Change gateway rssi type to signed int

+++ src/EnigmaIOTNode.h
@@ -92,7 +92,7 @@ typedef struct {
 	uint16_t nodeId; /**< Node identification */
 	uint8_t channel /*= DEFAULT_CHANNEL*/; /**< WiFi channel used on ESP-NOW communication */
 	uint8_t gateway[ENIGMAIOT_ADDR_LEN]; /**< Gateway address */
-	int8_t rssi; /**< Gateway signal strength */
+	signed int rssi; /**< Gateway signal strength */
 	uint8_t networkKey[KEY_LENGTH]; /**< Network key to protect key agreement */
 	char networkName[NETWORK_NAME_LENGTH]; /**< Network name. Used to search gateway peer */
 	bool sleepy; /**< Sleepy node */
